PlateauExperiment

Documentation for PlateauExperiment.

PlateauExperiment.IntegralSintOvert

PlateauExperiment.NoiseBall

PlateauExperiment.NoiseInterval

PlateauExperiment.T

PlateauExperiment.T_minus_one_one

PlateauExperiment.compute_residual

PlateauExperiment.convert_matrix

PlateauExperiment.symmetrize_density

PlateauExperiment.Υ

PlateauExperiment.τ_1

PlateauExperiment.τ_2

PlateauExperiment.IntegralSintOvert — Method

Compute a vector of K elements containing

\[\frac{1}{j\pi}\int_0^{j\pi} \frac{sin(t)}{t}dt\]

for $1 \leq j \leq K$ by using Taylor Models integration for $[\pi, j\pi]$.

PlateauExperiment.NoiseBall — Method

Build a Matrix of Balls representig Convolution with Gaussian Noise of variance σ Galerkin truncated at frequence K

PlateauExperiment.NoiseInterval — Method

Build a Matrix of Intervals representig Convolution with Gaussian Noise of variance σ Galerkin truncated at frequence K

PlateauExperiment.T — Method

Plateau map on [0, 1]

PlateauExperiment.T_minus_one_one — Method

Plateau map on [-1, 1]

PlateauExperiment.compute_residual — Method

Compute Pv-v

PlateauExperiment.convert_matrix — Method

Convert a matrix of Complex{Interval} to a BallMatrix with a center matrix of complex numbers

PlateauExperiment.symmetrize_density — Method

Symmetrize density

PlateauExperiment.Υ — Method

L2-norm of log(|T'|).

This can be computed explictly as

$\Upsilon(\alpha, \beta) = \sqrt{2}((\ln((1+\beta)\alpha)-(\alpha-1))^2+(\alpha-1)^2)^{\frac{1}{2}}$

PlateauExperiment.τ_1 — Method

Coordinate change from [-1,1] to [0,1]

PlateauExperiment.τ_2 — Method

Coordinate change from [0,1] to [-1,1]